Changed create group
