+++ budgetingapp/pages/Profile.js
@@ -96,7 +96,7 @@ export default function Profile({ navigation }) {
                 </Text>
             </TouchableOpacity>
             { /* FOR TESTING */ }
-            <TouchableOpacity style={styles.buttonOne} onPress={() => navigation.navigate("CreateGroup")}>
+            <TouchableOpacity style={styles.buttonOne} onPress={() => navigation.navigate("NoGroups")}>
                 <Text style={styles.buttonText}>Create Group</Text>
             </TouchableOpacity>
         </View>
